[storage] trivial: cleanup #[allow(unused_imports)]
Closes: #909
